% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_requirements.R
\name{capture_requirements}
\alias{capture_requirements}
\title{Capture packages in use}
\usage{
capture_requirements(
  file = "_requirements.txt",
  write = FALSE,
  attached = TRUE
)
}
\arguments{
\item{file}{A path to the filename for saving the \code{requirements.txt}. The default is \verb{_requirements.txt}.}

\item{write}{Boolean, \code{TRUE} or \code{FALSE}. If \code{TRUE} the function write the list of conda formatted dependencies in a file. If \code{FALSE}, it returns the dataframe with the package metadata.}

\item{attached}{Boolean, \code{TRUE} or \code{FALSE}. If \code{TRUE} the function returns only the attached packages. If \code{FALSE}, all the packages present in the workspace are included. Default is \code{TRUE}.}
}
\value{
A dataframe or a text file with conda formatted package metadata
}
\description{
This function creates a dependencies metadata \code{requirements.txt} file following the minimal conda compatible naming convention for package version definition:
\code{channel::package=version}
It can be used for creating a conda environment with \verb{conda create env --name tinyverse --file  requirements.txt}
}
\examples{
capture_requirements(write = FALSE, attached = TRUE)
}
